requirements.txt
playwright==1.48.0
pytest==8.3.3
pytest-playwright==0.5.0
pytest-xdist==3.6.1
pyyaml==6.0.2
python-dotenv==1.0.1
requests==2.32.3